refactor: 전체 코드 리팩터링
